Polish dashboard, separate out components
