+++ src/components/PromptGuide.svelte
@@ -1,0 +1,16 @@
+<div class="my-12 text-chalkboard-100 dark:text-chalkboard-20">
+	<h2 class="font-mono pt-1 text-xs uppercase">Prompt writing tips:</h2>
+	<ul class="list-disc pl-6 text-lg">
+		<li class="my-4">
+			Describe an object that can be represented in geometric shapes, and less nebulous concepts
+			such as "a tiger" or "the universe", unless you just want to see what it does with that 😏
+		</li>
+		<li class="my-4">
+			Be as explicit as possible. For example, if you want a plate with 4 holes, say <em>where</em>
+			you want the holes placed and <em>how big</em> of a diameter each should have
+		</li>
+		<li class="my-4">
+			Our ML models are better at describing single objects than assemblies <em>for now</em>
+		</li>
+	</ul>
+</div>
